refactor: restructure athlete management components and logic
