+++ src/components/data-table/data-table-filter-list.tsx
@@ -15,6 +15,7 @@ import type {
   FilterOperator,
   JoinOperator,
 } from '@/types/data-table';
+import type { DataTableControlledState } from '@/hooks/use-data-table';
 import { DataTableRangeFilter } from '@/components/data-table/data-table-range-filter';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
@@ -78,6 +79,7 @@ interface DataTableFilterListProps<TData> extends React.ComponentProps<
   typeof PopoverContent
 > {
   table: Table<TData>;
+  state?: DataTableControlledState;
   debounceMs?: number;
   throttleMs?: number;
   shallow?: boolean;
@@ -86,6 +88,7 @@ interface DataTableFilterListProps<TData> extends React.ComponentProps<
 
 export function DataTableFilterList<TData>({
   table,
+  state,
   debounceMs = DEBOUNCE_MS,
   throttleMs = THROTTLE_MS,
   shallow = true,
@@ -109,6 +112,7 @@ export function DataTableFilterList<TData>({
     getFiltersStateParser<TData>(columns.map((field) => field.id))
       .withDefault([])
       .withOptions({
+        ...state,
         clearOnDefault: true,
         shallow,
         throttleMs,
