+++ src/components/data-table/data-table-sort-list.tsx
@@ -7,6 +7,7 @@ import {
 import * as React from 'react';
 import type { ColumnSort, SortDirection, Table } from '@tanstack/react-table';
 
+import type { DataTableControlledState } from '@/hooks/use-data-table';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
 import {
@@ -46,11 +47,13 @@ interface DataTableSortListProps<TData> extends React.ComponentProps<
   typeof PopoverContent
 > {
   table: Table<TData>;
+  state?: DataTableControlledState;
   disabled?: boolean;
 }
 
 export function DataTableSortList<TData>({
   table,
+  state,
   disabled,
   ...props
 }: DataTableSortListProps<TData>) {
@@ -60,7 +63,7 @@ export function DataTableSortList<TData>({
   const [open, setOpen] = React.useState(false);
   const addButtonRef = React.useRef<HTMLButtonElement>(null);
 
-  const sorting = table.getState().sorting;
+  const sorting = state?.sorting ?? [];
   const onSortingChange = table.setSorting;
 
   const { columnLabels, columns } = React.useMemo(() => {
