+++ src/features/dashboard/athlete/hooks/use-athlete-manager-query.ts
@@ -1,0 +1,54 @@
+import {
+  parseAsArrayOf,
+  parseAsInteger,
+  parseAsString,
+  useQueryState,
+} from 'nuqs';
+import type { AthleteProfileData } from '@/features/dashboard/types';
+import { getSortingStateParser } from '@/lib/data-table/parsers';
+import { parseRangeParam } from '@/lib/data-table/utils';
+
+const SORTABLE_COLUMN_IDS = new Set([
+  'name',
+  'beltLevel',
+  'weight',
+  'affiliation',
+  'createdAt',
+]);
+
+export function useAthleteTableQuery() {
+  const [page] = useQueryState('page', parseAsInteger.withDefault(1));
+  const [perPage] = useQueryState('perPage', parseAsInteger.withDefault(10));
+  const [athleteCodeFilter] = useQueryState('athleteCode');
+  const [nameFilter] = useQueryState('name');
+  const [genderFilter] = useQueryState(
+    'gender',
+    parseAsArrayOf(parseAsString, ',')
+  );
+  const [affiliationFilter] = useQueryState('affiliation');
+  const [beltFilter] = useQueryState('beltLevel');
+  const [weightFilter] = useQueryState('weight');
+  const [sort] = useQueryState(
+    'sort',
+    getSortingStateParser<AthleteProfileData>(SORTABLE_COLUMN_IDS).withDefault(
+      []
+    )
+  );
+
+  const beltRange = parseRangeParam(beltFilter);
+  const weightRange = parseRangeParam(weightFilter);
+
+  return {
+    page,
+    perPage,
+    athleteCodeFilter,
+    nameFilter,
+    genderFilter,
+    affiliationFilter,
+    beltFilter,
+    weightFilter,
+    sort,
+    beltRange,
+    weightRange,
+  };
+}
